+++ data-manager/api/src/main/java/org/orbisgis/orbisdata/datamanager/api/dataset/IJdbcTableSummary.java
@@ -1,0 +1,7 @@
+package org.orbisgis.orbisdata.datamanager.api.dataset;
+
+/**
+ * Extension of the {@link ISummary} interface dedicated to the {@link IJdbcTable}.
+ */
+public interface IJdbcTableSummary extends ISummary {
+}