+++ data-manager/jdbc/src/test/java/org/orbisgis/orbisdata/datamanager/jdbc/dsl/BuilderResultTest.java
@@ -44,13 +44,10 @@ import org.h2gis.utilities.wrapper.SpatialResultSetImpl;
 import org.h2gis.utilities.wrapper.StatementWrapper;
 import org.junit.jupiter.api.Test;
 import org.orbisgis.commons.printer.Ascii;
+import org.orbisgis.orbisdata.datamanager.api.dataset.*;
 import org.orbisgis.orbisdata.datamanager.jdbc.JdbcDataSource;
 import org.orbisgis.orbisdata.datamanager.jdbc.JdbcSpatialTable;
 import org.orbisgis.orbisdata.datamanager.jdbc.TableLocation;
-import org.orbisgis.orbisdata.datamanager.api.dataset.DataBaseType;
-import org.orbisgis.orbisdata.datamanager.api.dataset.IDataSet;
-import org.orbisgis.orbisdata.datamanager.api.dataset.ISpatialTable;
-import org.orbisgis.orbisdata.datamanager.api.dataset.ITable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -164,6 +161,8 @@ public class BuilderResultTest {
             }
             return new SpatialResultSetImpl(resultSet, (StatementWrapper) getStatement());
         }
+
+        @Override public IJdbcTableSummary getSummary() { return null; }
     }
 
     /**
@@ -175,8 +174,8 @@ public class BuilderResultTest {
             super(parent, databaseType);
         }
 
-        @Override public ITable getTable(String tableName) {return null;}
-        @Override public ISpatialTable getSpatialTable(String tableName) {
+        @Override public IJdbcTable getTable(String tableName) {return null;}
+        @Override public IJdbcSpatialTable getSpatialTable(String tableName) {
             String name = TableLocation.parse(tableName, getDataBaseType().equals(DataBaseType.H2GIS)).toString( getDataBaseType().equals(DataBaseType.H2GIS));
             try {
                 if(!JDBCUtilities.tableExists(connection,name)){
@@ -200,7 +199,7 @@ public class BuilderResultTest {
             } catch (SQLException ignored) {}
             return null;
         }
-        @Override public IDataSet getDataSet(String name) {return null;}
+        @Override public IJdbcTable getDataSet(String name) {return null;}
         
         @Override
         public boolean hasTable(String tableName) {