+++ process-manager/process/src/test/java/org/orbisgis/orbisdata/processmanager/process/inoutput/InputOutputTest.java
@@ -38,8 +38,8 @@ package org.orbisgis.orbisdata.processmanager.process.inoutput;
 
 import groovy.lang.MissingMethodException;
 import org.junit.jupiter.api.Test;
-import org.orbisgis.orbisdata.processmanager.process.ProcessManager;
 import org.orbisgis.orbisdata.processmanager.api.IProcess;
+import org.orbisgis.orbisdata.processmanager.process.ProcessManager;
 
 import static org.junit.jupiter.api.Assertions.*;